fix: /multiplayer loading on chatbox click prop drilled properly
