+++ src/exports.js
@@ -15,6 +15,7 @@ import ChatBox from "./components/multiplayer/controlCenter/chatbox.jsx";
 import Login from "./components/login/login.jsx";
 import WordBox from "./components/connections/wordBox.jsx";
 import ControlCenter from "./components/connections/controlCenter.jsx";
+import MultiplayerControlCenter from "./components/multiplayer/controlCenter.jsx"
 import CategoryCard from "./components/create/categoryCard.jsx";
 import CreateControlCenter from "./components/create/controlCenter.jsx";
 import Mistakes from "./components/connections/mistakes.jsx";
@@ -25,6 +26,6 @@ import puzzle_fuzion from "./assets/puzzle_fusion.png"
 
 export { Home, Navbar, Wordle, MultiplayerWordle, MultiplayerInput, MultiplayerKeyboard,
     Connections, Create, Settings, Input, Keyboard, Loading, Multiplayer, ChatBox, Login,
-    WordBox, ControlCenter, CategoryCard, CreateControlCenter, Mistakes,
+    WordBox, ControlCenter, CategoryCard, CreateControlCenter, Mistakes, MultiplayerControlCenter,
     keyboard_light_mode_backspace, keyboard_dark_mode_backspace, puzzle_fuzion
 }
